refactor(transducer): 重构了transducer的目录结构
