+++ packages/screwdriver-js/src/transducer/Provider.ts
@@ -1,0 +1,3 @@
+export interface Provider<A> extends Function {
+  (): A;
+}
