+++ packages/screwdriver-js/src/transducer/Transducer.ts
@@ -1,0 +1,8 @@
+import { Mapper } from "..";
+import { Transformer } from "./Transformer";
+
+export interface Transducer<A, B, C> extends
+  Mapper<Transformer<B, C>, Transformer<A, C>>,
+  Function {
+  (xf: Transformer<B, C>): Transformer<A, C>;
+}
