+++ packages/screwdriver-js/src/transducer/Transformer.ts
@@ -1,0 +1,8 @@
+import { Reducer, Mapper } from "..";
+import { Provider } from "./Provider";
+
+export interface Transformer<A, C> {
+  '@@transducer/init': Provider<C>;
+  '@@transducer/step': Reducer<A, C>;
+  '@@transducer/result': Mapper<any, any>;
+}
